fix(plugin): contradiction-sync WASM-core failures on Node 26 (BigInt wrap + test fixture)
Test 21 (resolveWithCore vim-vs-vscode): production bug. The unified
_resolveWithCandidatesCore path passed Math.floor(nowUnixSeconds) (a JS
Number) to wasm_resolveWithCandidates, whose now_unix arg wasm-bindgen
declares as i64 -> BigInt. Node >=26 rejects the implicit Number->BigInt
coercion (Cannot convert 1777000000 to a BigInt) that older runtimes
tolerated; the throw was swallowed by the try/catch -> empty decisions
-> decisions[0] undefined -> reading .action crashed. Same family as the
X25519 Node-26 fix. Wraps with BigInt() to match the legacy path
(contradiction-sync.ts:621) and loadWeightsFile (:250). This was breaking
ALL contradiction resolution on Node 26, not just the test.

Test 12 (isPinnedClaim st=p -> true): test-only fixture bug, unrelated to
Node 26. The bare {t,c,st} object is missing the v0 Claim's required
fields (cf/i/sa), so serde_json::from_str::<Claim> fails and the WASM
core's is_pinned_json returns false before the st=='p' check runs. Uses
makeClaim() so the fixture carries the full v0 field set.

+++ skill/plugin/contradiction-sync.ts
@@ -514,7 +514,10 @@ function _resolveWithCandidatesCore(
       weightsJson,
       thresholdLower,
       thresholdUpper,
-      Math.floor(nowUnixSeconds),
+      // wasm-bindgen declares `now_unix` as i64 → JS BigInt. Node ≥26 rejects
+      // the implicit Number→BigInt coercion that older runtimes tolerated, so
+      // wrap explicitly (same as the legacy path below + loadWeightsFile).
+      BigInt(Math.floor(nowUnixSeconds)),
       TIE_ZONE_SCORE_TOLERANCE,
     );
   } catch (err) {
